Rewrite approximate() and num_deltas(), etc.
The approximate() does not discard the zeroed highest-rank delta, but
instead returns the number of such zeros. A '0' is returned when all
the recorded values are independent. In case of failure, the object is
left intact.

The num_deltas() is extended to allow "actual" derivatives counting. The
reduce() can be used to cut by rank only, but not by minimal value.

Minor changes in local variable names and comments.