Split object at a shock point functionality
The split_at_gap() is to split an object into two: before and after the
"shock". The split-point is specified as a rank, returned by find_gap().

The align_times() is to align time-intervals between two objects, to
allow adding or subtracting. The python-style negative indices can be
passed to get_value_time().

Add 'keep_time' argument to extrapolate(), similarly to reduce().